requirements.txt
appdirs==1.4.4
asgiref==3.3.1
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
commonmark==0.9.1
commonmarkextensions==0.0.6
cssutils==2.3.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-debug-toolbar==3.2
django-heroku==0.3.1
django-html-emailer==0.1.0
django-phonenumber-field==5.0.0
django-picklefield==3.0.1
Django-Verify-Email==1.0.9
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.3
filelock==3.0.12
idna==3.3
inlinestyler==0.2.5
lxml==4.7.1
packaging==21.3
phonenumbers==8.12.20
pipenv==2021.11.23
pipreqs==0.4.11
platformdirs==2.4.0
pluggy==1.0.0
psycopg2==2.9.2
py==1.11.0
pycodestyle==2.6.0
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tox==3.23.0
urllib3==1.26.7
validate-email==1.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
yarg==0.1.9
